reduce redundant files
